refactor file architecture
