+++ src/components/display/OneEvent.js
@@ -4,7 +4,7 @@ import { useHistory } from 'react-router-dom';
 import { useQuery, useMutation } from '@apollo/react-hooks';
 import { GET_ONE_EVENT, GET_ONE_DATE } from '../../graphQl/queries';
 import { DELETE_EVENT} from '../../graphQl/mutations/eventM'
-import NavBar from './NavBar';
+import NavBar from './Nav/Index';
 // Styling imports
 import { Card, Typography, makeStyles, Button } from '@material-ui/core';
 import ListItemText from '@material-ui/core/ListItemText';
